# Network data state management modifications