--- a/config/yaml_config.go
+++ b/config/yaml_config.go
@@ -259,6 +259,10 @@ func mergeNetworkYamlConfig(agentConf *AgentConfig, networkConf *YamlAgentConfig
 		}
 	}
 
+	// Pull additional parameters from the global config file.
+	agentConf.LogLevel = ddconfig.Datadog.GetString("log_level")
+	agentConf.StatsdPort = ddconfig.Datadog.GetInt("dogstatsd_port")
+
 	return agentConf, nil
 }