--- a/ebpf/state.go
+++ b/ebpf/state.go
@@ -15,9 +15,9 @@ const (
 	DEBUGCLIENT = "-1"
 
 	// defaultMaxClosedConns & defaultMaxClientStats are the maximum number of objects that can be stored in-memory.
-	// With clients checking connection stats roughly every 30s, this gives us roughly 2500 + 5000 objects a second respectively.
+	// With clients checking connection stats roughly every 30s, this gives us roughly 2.5k + ~3k objects a second respectively.
 	defaultMaxClosedConns = 75000 // ~100 bytes per conn = 7.5MB
-	defaultMaxClientStats = 150000
+	defaultMaxClientStats = 100000
 	defaultExpiry         = 2 * time.Minute
 	defaultClientInterval = 30 * time.Second
 )