- Getting rid of our `override` stats concept as it's difficult to debug race conditions, simplifying the state code by a lot.
- Dealing with a lot of unexpected edge cases (out of order messages, underflows) by emitting metrics so we can use them to help debug problems.
- Adding closed connection map + stat counter size restrictions so we don't run away with memory when triggering unhandled race conditions.
- Monotonic counters for connections that close/re-open on the same IP:port mappings, over interval bounds will be reset.

**Before**
```
BenchmarkStoreClosedConnection/StoreClosedConnection-100         	   20000	     69614 ns/op	   18401 B/op	     599 allocs/op
BenchmarkStoreClosedConnection/StoreClosedConnection-1000        	    2000	   1114043 ns/op	  184181 B/op	    5999 allocs/op
BenchmarkStoreClosedConnection/StoreClosedConnection-10000       	      50	  22068868 ns/op	 1896434 B/op	   59811 allocs/op
BenchmarkStoreClosedConnection/StoreClosedConnection-30000       	      50	  37215887 ns/op	 5751271 B/op	  179445 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-100-0                     	   20000	     93261 ns/op	   34365 B/op	     412 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-100-50                    	   20000	     69812 ns/op	   34367 B/op	     412 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-100-100                   	   20000	     72119 ns/op	   34368 B/op	     412 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-1000-0                    	    2000	    664163 ns/op	  310069 B/op	    4015 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-1000-500                  	    2000	    634678 ns/op	  310269 B/op	    4016 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-1000-1000                 	    2000	    685621 ns/op	  310426 B/op	    4017 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-10000-0                   	     100	  10077985 ns/op	 5498000 B/op	   40127 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-10000-5000                	     100	  10424169 ns/op	 5536285 B/op	   40328 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-10000-10000               	     100	  11805941 ns/op	 5589957 B/op	   40529 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-30000-0                   	      50	  37313849 ns/op	18039313 B/op	  120653 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-30000-15000               	      30	  35325675 ns/op	18509785 B/op	  123083 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-30000-30000               	      30	  40856162 ns/op	18824539 B/op	  125103 allocs/op
PASS
ok  	command-line-arguments	34.152s
```

**After**
```
goos: linux
goarch: amd64
BenchmarkStoreClosedConnection/StoreClosedConnection-100         	   50000	     28173 ns/op	    2400 B/op	     300 allocs/op
BenchmarkStoreClosedConnection/StoreClosedConnection-1000        	    5000	    295882 ns/op	   24114 B/op	    3000 allocs/op
BenchmarkStoreClosedConnection/StoreClosedConnection-10000       	     500	   2855677 ns/op	  249219 B/op	   30020 allocs/op
BenchmarkStoreClosedConnection/StoreClosedConnection-30000       	     200	   9675471 ns/op	  810351 B/op	   90155 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-100-0                     	   30000	     45052 ns/op	   34365 B/op	     412 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-100-50                    	   30000	     45742 ns/op	   34366 B/op	     412 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-100-100                   	   30000	     45043 ns/op	   34366 B/op	     412 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-1000-0                    	    3000	    442298 ns/op	  310041 B/op	    4015 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-1000-500                  	    3000	    427614 ns/op	  310133 B/op	    4015 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-1000-1000                 	    3000	    439425 ns/op	  310198 B/op	    4015 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-10000-0                   	     200	   6931127 ns/op	 5490722 B/op	   40076 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-10000-5000                	     200	   6740125 ns/op	 5504340 B/op	   40076 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-10000-10000               	     200	   7048974 ns/op	 5525702 B/op	   40076 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-30000-0                   	      50	  26199812 ns/op	18039300 B/op	  120653 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-30000-15000               	      50	  27069750 ns/op	18206788 B/op	  120653 allocs/op
BenchmarkConnectionsGet/ConnectionsGet-30000-30000               	      50	  28328642 ns/op	18322880 B/op	  120652 allocs/op
PASS
ok  	command-line-arguments	30.286s
```