@kevinconaway 2019-02-26
What is the default log level in the agent?  There are a handful of messages at the WARN level that we might to show up and may not if the default is ERROR